Moved output outside the main renderer
