+++ src/renderer/common/display/icon-list-item.js
