+++ src/renderer/common/display/icon-text.js
