+++ src/renderer/common/display/simple-text.js
