+++ src/renderer/common/output-list.css
@@ -1,0 +1,16 @@
+
+#listbox::-webkit-scrollbar {
+    width: 0.25rem;
+}
+
+#listbox::-webkit-scrollbar-track,
+#listbox::-webkit-scrollbar-track-piece,
+#listbox::-webkit-resizer,
+#listbox::-webkit-scrollbar-corner,
+#listbox::-webkit-scrollbar-button {
+    display: none;
+}
+
+#listbox::-webkit-scrollbar-thumb {
+    background: lightgrey;
+}
