+++ src/renderer/main/index.css
@@ -15,22 +15,6 @@ html, body, * {
     background: transparent;
 }
 
-#listbox::-webkit-scrollbar {
-    width: 0.25rem;
-}
-
-#listbox::-webkit-scrollbar-track,
-#listbox::-webkit-scrollbar-track-piece,
-#listbox::-webkit-resizer,
-#listbox::-webkit-scrollbar-corner,
-#listbox::-webkit-scrollbar-button {
-    display: none;
-}
-
-#listbox::-webkit-scrollbar-thumb {
-    background: lightgrey;
-}
-
 .MuiOutlinedInput-notchedOutline {
     border: none !important;
 }
